requirements.txt
anyio==3.7.0
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
decorator==4.4.2
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
mysqlclient==2.0.2
pycparser==2.20
python-dotenv==0.15.0
python-telegram-bot==20.3
pytz==2020.4
six==1.15.0
sniffio==1.3.0
SQLAlchemy==1.3.22
tornado==6.1
tzlocal==2.1
